FastTransforms.jl Documentation

Fast Transforms

FastTransforms.cjt — Function.

Computes the Chebyshev expansion coefficients given the Jacobi expansion coefficients $c$ with parameters $α$ and $β$.

See also icjt and jjt.

source

FastTransforms.icjt — Function.

Computes the Jacobi expansion coefficients with parameters $α$ and $β$ given the Chebyshev expansion coefficients $c$.

See also cjt and jjt.

source

FastTransforms.jjt — Function.

Computes the Jacobi expansion coefficients with parameters $γ$ and $δ$ given the Jacobi expansion coefficients $c$ with parameters $α$ and $β$.

See also cjt and icjt.

source

FastTransforms.plan_cjt — Function.

Pre-plan optimized DCT-I and DST-I plans and pre-allocate the necessary arrays, normalization constants, and recurrence coefficients for a forward Chebyshev—Jacobi transform.

$c$ is the vector of coefficients; and,

$α$ and $β$ are the Jacobi parameters.

Optionally:

$M$ determines the number of terms in Hahn's asymptotic expansion.

source

FastTransforms.plan_icjt — Function.

Pre-plan optimized DCT-I and DST-I plans and pre-allocate the necessary arrays, normalization constants, and recurrence coefficients for an inverse Chebyshev—Jacobi transform.

$c$ is the vector of coefficients; and,

$α$ and $β$ are the Jacobi parameters.

Optionally:

$M$ determines the number of terms in Hahn's asymptotic expansion.

source

FastTransforms.paduatransform — Function.

Padua Transform maps from interpolant values at the Padua points to the 2D Chebyshev coefficients.

source

FastTransforms.ipaduatransform — Function.

Inverse Padua Transform maps the 2D Chebyshev coefficients to the values of the interpolation polynomial at the Padua points.

source

FastTransforms.plan_paduatransform! — Function.

Pre-plan a Padua Transform.

source

FastTransforms.plan_ipaduatransform! — Function.

Pre-plan an Inverse Padua Transform.

source

Other Exported Methods

FastTransforms.gaunt — Function.

Calculates the Gaunt coefficients, defined by:

a(m,n,μ,ν,q) = (2(n+ν-2q)+1)/2 (n+ν-2q-m-μ)!/(n+ν-2q+m+μ)! ∫₋₁⁺¹ P_m^n(x) P_ν^μ(x) P_{n+ν-2q}^{m+μ}(x) dx.

or defined by:

P_n^m(x) P_ν^μ(x) = ∑_{q=0}^{q_{max}} a(m,n,μ,ν,q) P_{n+ν-2q}^{m+μ}(x)

This is a Julia implementation of the stable recurrence described in:

Y.-l. Xu, "Fast evaluation of Gaunt coefficients: recursive approach", J. Comp. Appl. Math., 85:53–65, 1997.

source

Calculates the Gaunt coefficients in 64-bit floating-point arithmetic.

source

Internal Methods

clenshawcurtis

clenshawcurtisweights

fejer1

fejer2

fejerweights1

fejerweights2

FastTransforms.paduapoints — Function.

Returns coordinates of the (n+1)(n+2)/2 Padua points.

source

half

two

δ

Λ

pochhammer

stirlingseries