Home

FastTransforms.jl Documentation

Introduction

In numerical analysis, it is customary to expand a function in a basis:

\[f(x) \sim \sum_{\ell=0}^{\infty} f_{\ell} \phi_{\ell}(x).\]

It may be necessary to transform our representation to one in a new basis, say, $\{\psi_m(x)\}_{m\ge0}$:

\[f(x) \sim \sum_{m=0}^{\infty} g_m \psi_m(x).\]

In many cases of interest, both representations are of finite length $n$ and we seek a fast method (faster than $\mathcal{O}(n^2)$) to transform the original coefficients $f_{\ell}$ to the new coefficients $g_m$.

A similar problem arises when we wish to evaluate $f$ at a set of points $\{x_m\}_{m=0}^n$. We wish to transform coefficients of $f$ to values at the set of points in fewer than $\mathcal{O}(n^2)$ operations.

Fast Transforms

FastTransforms.leg2cheb — Function.

Computes the Chebyshev expansion coefficients given the Legendre expansion coefficients:

\[{\rm CLT} : \sum_{n=0}^N c_n^{\rm leg}P_n(x) \to \sum_{n=0}^N c_n^{\rm cheb}T_n(x).\]

source

FastTransforms.cheb2leg — Function.

Computes the Legendre expansion coefficients given the Chebyshev expansion coefficients:

\[{\rm iCLT} : \sum_{n=0}^N c_n^{\rm cheb}T_n(x) \to \sum_{n=0}^N c_n^{\rm leg}P_n(x).\]

source

FastTransforms.plan_leg2cheb — Function.

Pre-computes the Legendre–Chebyshev transform.

source

FastTransforms.plan_cheb2leg — Function.

Pre-computes the Chebyshev–Legendre transform.

source

FastTransforms.cjt — Function.

Computes the Chebyshev expansion coefficients given the Jacobi expansion coefficients $c$ with parameters $\alpha$ and $\beta$:

\[{\rm CJT} : \sum_{n=0}^N c_n^{\rm jac}P_n^{(\alpha,\beta)}(x) \to \sum_{n=0}^N c_n^{\rm cheb}T_n(x).\]

source

FastTransforms.icjt — Function.

Computes the Jacobi expansion coefficients with parameters $\alpha$ and $\beta$ given the Chebyshev expansion coefficients $c$:

\[{\rm iCJT} : \sum_{n=0}^N c_n^{\rm cheb}T_n(x) \to \sum_{n=0}^N c_n^{\rm jac}P_n^{(\alpha,\beta)}(x).\]

source

FastTransforms.jjt — Function.

Computes the Jacobi expansion coefficients with parameters $\gamma$ and $\delta$ given the Jacobi expansion coefficients $c$ with parameters $\alpha$ and $\beta$:

\[{\rm JJT} : \sum_{n=0}^N c_n^{\rm jac}P_n^{(\alpha,\beta)}(x) \to \sum_{n=0}^N c_n^{\rm jac}P_n^{(\gamma,\delta)}(x).\]

source

FastTransforms.plan_cjt — Function.

Pre-plan optimized DCT-I and DST-I plans and pre-allocate the necessary arrays, normalization constants, and recurrence coefficients for a forward Chebyshev—Jacobi transform.

$c$ is the vector of coefficients; and,

$\alpha$ and $\beta$ are the Jacobi parameters.

Optionally:

$M$ determines the number of terms in Hahn's asymptotic expansion.

source

FastTransforms.plan_icjt — Function.

Pre-plan optimized DCT-I and DST-I plans and pre-allocate the necessary arrays, normalization constants, and recurrence coefficients for an inverse Chebyshev—Jacobi transform.

$c$ is the vector of coefficients; and,

$\alpha$ and $\beta$ are the Jacobi parameters.

Optionally:

$M$ determines the number of terms in Hahn's asymptotic expansion.

source

FastTransforms.paduatransform — Function.

Padua Transform maps from interpolant values at the Padua points to the 2D Chebyshev coefficients.

source

FastTransforms.ipaduatransform — Function.

Inverse Padua Transform maps the 2D Chebyshev coefficients to the values of the interpolation polynomial at the Padua points.

source

FastTransforms.plan_paduatransform! — Function.

Pre-plan a Padua Transform.

source

FastTransforms.plan_ipaduatransform! — Function.

Pre-plan an Inverse Padua Transform.

source

FastTransforms.sph2fourier — Function.

Computes the bivariate Fourier series given by the spherical harmonic expansion:

\[{\rm SHT} : \sum_{\ell=0}^n\sum_{m=-\ell}^{\ell} f_{\ell}^m Y_{\ell}^m(\theta,\varphi) \to \sum_{\ell=0}^n\sum_{m=-n}^{n} g_{\ell}^m \frac{e^{{\rm i} m \varphi}}{\sqrt{2\pi}} \left\{\begin{array}{c}\cos\ell\theta\\ \sin(\ell+1)\theta\end{array}\right\},\]

where the cosines are used when $m$ is even and the sines are used when $m$ is odd. The spherical harmonic expansion coefficients are organized as follows:

\[F = \begin{pmatrix} f_0^0 & f_1^{-1} & f_1^1 & f_2^{-2} & f_2^2 & \cdots & f_n^{-n} & f_n^n\\ f_1^0 & f_2^{-1} & f_2^1 & f_3^{-2} & f_3^2 & \cdots & 0 & 0\\ \vdots & \vdots & \vdots & \vdots & \vdots & \ddots & \vdots & \vdots\\ f_{n-2}^0 & f_{n-1}^{-1} & f_{n-1}^1 & f_n^{-2} & f_n^2 & & \vdots & \vdots\\ f_{n-1}^0 & f_n^{-1} & f_n^1 & 0 & 0 & \cdots & 0 & 0\\ f_n^0 & 0 & 0 & 0 & 0 & \cdots & 0 & 0\\ \end{pmatrix},\]

and the Fourier coefficients are organized similarly:

\[G = \begin{pmatrix} g_0^0 & g_0^{-1} & g_0^1 & \cdots & g_0^{-n} & g_0^n\\ g_1^0 & g_1^{-1} & g_1^1 & \cdots & g_1^{-n} & g_1^n\\ \vdots & \vdots & \vdots & \ddots & \vdots & \vdots\\ g_{n-1}^0 & g_{n-1}^{-1} & g_{n-1}^1& \cdots & g_{n-1}^{-n} & g_{n-1}^n\\ g_n^0 & 0 & 0 & \cdots & g_n^{-n} & g_n^n\\ \end{pmatrix}.\]

source

FastTransforms.fourier2sph — Function.

Computes the spherical harmonic expansion given by the bivariate Fourier series:

\[{\rm iSHT} : \sum_{\ell=0}^n\sum_{m=-n}^{n} g_{\ell}^m \frac{e^{{\rm i} m \varphi}}{\sqrt{2\pi}} \left\{\begin{array}{c}\cos\ell\theta\\ \sin(\ell+1)\theta\end{array}\right\} \to \sum_{\ell=0}^n\sum_{m=-\ell}^{\ell} f_{\ell}^m Y_{\ell}^m(\theta,\varphi),\]

where the cosines are used when $m$ is even and the sines are used when $m$ is odd. The spherical harmonic expansion coefficients are organized as follows:

\[F = \begin{pmatrix} f_0^0 & f_1^{-1} & f_1^1 & f_2^{-2} & f_2^2 & \cdots & f_n^{-n} & f_n^n\\ f_1^0 & f_2^{-1} & f_2^1 & f_3^{-2} & f_3^2 & \cdots & 0 & 0\\ \vdots & \vdots & \vdots & \vdots & \vdots & \ddots & \vdots & \vdots\\ f_{n-2}^0 & f_{n-1}^{-1} & f_{n-1}^1 & f_n^{-2} & f_n^2 & & \vdots & \vdots\\ f_{n-1}^0 & f_n^{-1} & f_n^1 & 0 & 0 & \cdots & 0 & 0\\ f_n^0 & 0 & 0 & 0 & 0 & \cdots & 0 & 0\\ \end{pmatrix},\]

and the Fourier coefficients are organized similarly:

\[G = \begin{pmatrix} g_0^0 & g_0^{-1} & g_0^1 & \cdots & g_0^{-n} & g_0^n\\ g_1^0 & g_1^{-1} & g_1^1 & \cdots & g_1^{-n} & g_1^n\\ \vdots & \vdots & \vdots & \ddots & \vdots & \vdots\\ g_{n-1}^0 & g_{n-1}^{-1} & g_{n-1}^1& \cdots & g_{n-1}^{-n} & g_{n-1}^n\\ g_n^0 & 0 & 0 & \cdots & g_n^{-n} & g_n^n\\ \end{pmatrix}.\]

source

FastTransforms.plan_sph2fourier — Function.

Pre-computes the spherical harmonic transform.

source

Other Exported Methods

FastTransforms.gaunt — Function.

Calculates the Gaunt coefficients, defined by:

\[a(m,n,\mu,\nu,q) = \frac{2(n+\nu-2q)+1}{2} \frac{(n+\nu-2q-m-\mu)!}{(n+\nu-2q+m+\mu)!} \int_{-1}^{+1} P_m^n(x) P_\nu^\mu(x) P_{n+\nu-2q}^{m+\mu}(x) {\rm\,d}x.\]

or defined by:

\[P_n^m(x) P_\nu^\mu(x) = \sum_{q=0}^{q_{\rm max}} a(m,n,\mu,\nu,q) P_{n+\nu-2q}^{m+\mu}(x)\]

This is a Julia implementation of the stable recurrence described in:

Y.-l. Xu, Fast evaluation of Gaunt coefficients: recursive approach, J. Comp. Appl. Math., 85:53–65, 1997.

source

Calculates the Gaunt coefficients in 64-bit floating-point arithmetic.

source

FastTransforms.paduapoints — Function.

Returns coordinates of the $(n+1)(n+2)/2$ Padua points.

source

FastTransforms.sphevaluate — Function.

Pointwise evaluation of spherical harmonic $Y_{\ell}^m(\theta,\varphi)$.

source

Internal Methods

FastTransforms.half — Function.

Compute a typed 0.5.

source

FastTransforms.two — Function.

Compute a typed 2.

source

FastTransforms.δ — Function.

The Kronecker $\delta$ function:

\[\delta_{k,j} = \left\{\begin{array}{ccc} 1 & {\rm for} & k = j,\\ 0 & {\rm for} & k \ne j.\end{array}\right.\]

source

FastTransforms.Λ — Function.

The Lambda function $\Lambda(z) = \frac{\Gamma(z+\frac{1}{2})}{\Gamma(z+1)}$ for the ratio of gamma functions.

source

For 64-bit floating-point arithmetic, the Lambda function uses the asymptotic series for $\tau$ in Appendix B of

I. Bogaert and B. Michiels and J. Fostier, 𝒪(1) computation of Legendre polynomials and Gauss–Legendre nodes and weights for parallel computing, SIAM J. Sci. Comput., 34:C83–C101, 2012.

source

The Lambda function $\Lambda(z,λ₁,λ₂) = \frac{\Gamma(z+\lambda_1)}{Γ(z+\lambda_2)}$ for the ratio of gamma functions.

source

FastTransforms.pochhammer — Function.

Pochhammer symbol $(x)_n = \frac{\Gamma(x+n)}{\Gamma(x)}$ for the rising factorial.

source

FastTransforms.stirlingseries — Function.

Stirling's asymptotic series for $\Gamma(z)$.

source

FastTransforms.clenshawcurtis — Function.

Compute nodes and weights of the Clenshaw—Curtis quadrature rule with a Jacobi weight.

source

FastTransforms.clenshawcurtisweights — Function.

Compute weights of the Clenshaw—Curtis quadrature rule with a Jacobi weight.

source

FastTransforms.fejer1 — Function.

Compute nodes and weights of Fejer's first quadrature rule with a Jacobi weight.

source

FastTransforms.fejer2 — Function.

Compute nodes and weights of Fejer's second quadrature rule with a Jacobi weight.

source

FastTransforms.fejerweights1 — Function.

Compute weights of Fejer's first quadrature rule with a Jacobi weight.

source

FastTransforms.fejerweights2 — Function.

Compute weights of Fejer's second quadrature rule with a Jacobi weight.

source

FastTransforms.chebyshevjacobimoments1 — Function.

Modified Chebyshev moments of the first kind with respect to the Jacobi weight:

\[ \int_{-1}^{+1} T_n(x) (1-x)^\alpha(1+x)^\beta{\rm\,d}x.\]

source

FastTransforms.chebyshevjacobimoments2 — Function.

Modified Chebyshev moments of the second kind with respect to the Jacobi weight:

\[ \int_{-1}^{+1} U_n(x) (1-x)^\alpha(1+x)^\beta{\rm\,d}x.\]

source

FastTransforms.incrementα! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha+1,\beta)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\beta)}(x)$ in-place.

source

FastTransforms.incrementβ! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha,\beta+1)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\beta)}(x)$ in-place.

source

FastTransforms.incrementαβ! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha+1,\alpha+1)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\alpha)}(x)$ in-place.

source

FastTransforms.decrementα! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha-1,\beta)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\beta)}(x)$ in-place.

source

FastTransforms.decrementβ! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha,\beta-1)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\beta)}(x)$ in-place.

source

FastTransforms.decrementαβ! — Function.

Compute Jacobi expansion coefficients in $P_n^{(\alpha-1,\alpha-1)}(x)$ given Jacobi expansion coefficients in $P_n^{(\alpha,\alpha)}(x)$ in-place.

source